Add a couple of additional tests, rename a couple of tests to correct names
